+++ src/web-client/constants.py
@@ -9,4 +9,4 @@ LOGIN_ENDPOINT = "/login"
 GET_USER_ENDPOINT = "/get_current_user"
 GET_PUBLIC_KEY_ENDPOINT = "/get_public_key"
 RESET_PASSWORD_ENDPOINT = "/change_password"
-GET_FILES = "/api/files"
+FILES_ENDPOINT = "/api/files"
