+++ src/web-client/requirements.txt
@@ -1,5 +1,8 @@
 Flask==3.0.2
 Werkzeug==3.0.1
-requests 
+requests==2.32.3
 PyNaCl==1.5.0  
-pytest
+pytest==8.3.5
+pydantic==2.11.4
+pydantic_core==2.33.2
+pydantic-settings==2.9.1
