+++ src/web-client/utils/encryption_helper.py
